FND-009 - Warm the browser under test, not a hardcoded chromium
Review round 3 approved the previous commit but flagged that its warm-up
never ran in the firefox half of the CI matrix. It was right, and the
first repair I reached for was also wrong.

`.github/workflows/ci.yml` installs only the matrix browser
(`playwright install --with-deps ${{ matrix.browser }}`), so a
`globalSetup` calling `chromium.launch()` has no binary in the firefox
job. It logged a launch failure, continued — by design, it is non-fatal —
and left that half of the matrix cold while docs/testing.md claimed
belt-and-braces coverage for both.

Reading the browser from `config.projects[0].use.defaultBrowserType` does
not fix it: `FullConfig.projects` is NOT filtered by `--project`. Verified
rather than assumed — put firefox first in a throwaway config, ran
`--project=chromium`, and globalSetup still reported
`projects[0] === firefox`. So globalSetup cannot know which browser the
run selected, and any fix built on that is a fix in appearance only.

Use a Playwright setup project instead. Playwright resolves it per
project, so `warmup:chromium`/`warmup:firefox` each carry their own
browser, `--project=firefox` pulls in `warmup:firefox` with no env
plumbing, and the warm-up's outcome lands in the report instead of only
stdout.

It still cannot fail the suite. A setup project's failure normally blocks
its dependents, which would be wrong here: `optimizeDeps.include` is the
load-bearing fix and this is a backstop for whatever is left off that
list. Measured across 11 cold runs with no effective warm-up, one failed;
across 4 cold runs with it, none did. Its absence worsens the odds rather
than guaranteeing a failure, so it must not be able to turn a warm-up
hiccup into a red suite. `slice.spec.ts` still makes every assertion
itself, so this cannot turn a real failure green either.

Evidence, against a deliberately cleared `node_modules/.vinxi`:

- `--project=firefox --list` schedules `warmup:firefox`, `--project=chromium`
  schedules `warmup:chromium` — the bug this commit exists to fix.
- emulator browser E2E (chromium, cold): 2 passed, warm-up 16.3s +
  slice 5.0s, and it logs "editor reached in chromium" rather than the
  silent launch failure it used to.
- typecheck and check:ci pass.

docs/testing.md now records why both obvious alternatives are broken, so
the next person does not re-derive it, and stops claiming coverage the
firefox job did not have.

Firefox itself is still unverified locally — only Chromium exists in this
container — so that matrix job remains first-run in CI.
