+++ docs/testing.md
@@ -87,10 +87,16 @@ Vite does not pre-bundle a dependency until something imports it, and each disco
 
 CI is always the cold case — a fresh checkout has no `node_modules/.vinxi`. `retries: 2` would usually have hidden this (the dev server survives between retries, so retry #1 sees a warm cache), which is worse than failing: the suite goes green and the real cause stays invisible.
 
-Two mechanisms, deliberately belt-and-braces:
+Two mechanisms, and they are **not** equal partners — the first is load-bearing and the second is a backstop:
 
-- **`app.config.ts`'s `optimizeDeps.include`** pre-bundles those dependencies at dev-server start, collapsing all four rounds into one startup cost. This also removes the stutter from plain `bun run dev`. It is an optimization, not a contract — a dependency added later and left off the list still works, it just reintroduces one reload for itself. `tone` is the easy one to miss, because nothing on the dashboard imports it, so it is only discovered when a project page first mounts.
-- **`e2e-emulator/warmDevServer.ts`** (Playwright `globalSetup`, runs after `webServer` is up) walks dashboard → new project → editor once in a real browser, so any remaining optimize-and-reload happens before the first test. It must reach the *editor*, not just the dashboard, for the `tone` reason above. It never fails the run, and it logs precisely whether it confirmed a warm editor rather than claiming success it did not achieve.
+- **`app.config.ts`'s `optimizeDeps.include`** pre-bundles those dependencies at dev-server start, collapsing all four rounds into one startup cost. This also removes the stutter from plain `bun run dev`. Measured: zero `new dependencies optimized` messages across four cold runs with it, versus a deterministic reproduction of the failure without it. It is an optimization, not a contract — a dependency added later and left off the list still works, it just reintroduces one reload for itself. `tone` is the easy one to miss, because nothing on the dashboard imports it, so it is only discovered when a project page first mounts. (`solid-firebase` is deliberately absent: the dashboard imports it during initial load, so it lands in Vite's first pre-load scan rather than a mid-session discovery round.)
+- **`e2e-emulator/warmDevServer.setup.ts`** walks dashboard → new project → editor once, so any remaining optimize-and-reload happens before the first assertion. It must reach the *editor*, not just the dashboard, for the `tone` reason above. It never fails the suite, and it logs precisely whether it confirmed a warm editor rather than claiming success it did not achieve.
+
+  It runs as a Playwright **setup project**, not a `globalSetup`, and the reason is worth recording because both obvious alternatives are broken. `.github/workflows/ci.yml` installs only the matrix browser (`playwright install --with-deps ${{ matrix.browser }}`), so a `globalSetup` calling `chromium.launch()` directly has no binary in the firefox job and silently no-ops there. Reading the browser from `config.projects[0].use.defaultBrowserType` does not rescue it either: **`FullConfig.projects` is not filtered by `--project`** — verified by putting firefox first in a throwaway config and running `--project=chromium`, which still reported `projects[0] === firefox`. A setup project resolves per project, so `warmup:chromium`/`warmup:firefox` each carry their own browser, `--project=firefox` pulls in `warmup:firefox` automatically, and the warm-up's outcome appears in the report rather than only in stdout.
+
+How much the backstop actually buys: across 11 cold runs with no effective warm-up, one failed; across 4 cold runs with it, none did. Its absence worsens the odds rather than guaranteeing a failure, which is why it stays non-fatal — and why `optimizeDeps.include` is the mechanism to keep correct.
+
+Creating a project in the warm-up cannot disturb `slice.spec.ts`'s `No projects yet` precondition: `listProjects` scopes on `where("ownerId", "==", ownerId)`, and `globalSetup`'s browser is a fresh profile with its own anonymous identity, as is each test's context.
 
 Note for anyone extending this suite: do not wait on `networkidle`. The app holds an open Firestore listener, so the network never goes idle and the wait can only time out. Wait for real elements — those locators re-resolve across a reload, which is the behaviour you want.
